Project Management Automation: Reorganize folder structure
