+++ packages/project-management-automation/lib/tasks/assign-fixed-issues/README.md
@@ -1,0 +1,10 @@
+Assign Fixed Issues
+===
+
+Adds assignee for issues which are marked to be "Fixed" by a pull request, and adds the "In Progress" label.
+
+## Rationale
+
+For users who would be affected by an issue (including the issue's creator), it is useful to receive progress updates once efforts have begun to address the issue.
+
+[Issue assignment](https://help.github.com/en/github/managing-your-work-on-github/assigning-issues-and-pull-requests-to-other-github-users) is a useful GitHub feature for developers, as it can avoid a scenario where multiple developers spend effort toward resolving the same issue due to the fact that progress was not clearly communicated.
