+++ packages/project-management-automation/lib/tasks/first-time-contributor/README.md
@@ -1,0 +1,12 @@
+First Time Contributor
+===
+
+Adds the "First Time Contributor" label to pull requests merged on behalf of contributors that have not previously made a contribution, and prompts the user to link their GitHub account to their WordPress.org profile if necessary for release notes credit.
+
+## Rationale
+
+A person's first contribution to the project is an exciting moment worth celebrating.
+
+Rate of new contributorship can be a potential metric of a project's open-source health.
+
+Compiling credits for release notes of a WordPress release [requires that a person's WordPress.org profile can be associated with a GitHub account](https://make.wordpress.org/core/2020/03/19/associating-github-accounts-with-wordpress-org-profiles/). This is not an obvious requirement for new contributors, and it should be brought to their attention at the earliest opportunity in their contributing journey.
